Rewrite sending websocket messages
Before:
One can send messages to all users in a room, but have to provide an user id to be skipped.

Now:
Split in send_message_all without an id to be skipped and send_message_skip_user which is
exactly the previous send_message.

Also:
Implemented send_message_user to send a message to a specific user. Currently dead code,
will be removen whenever the function is used.