Implement receiving all room info on join
On joining a room, the user now receives a message with type 'all' containing
all users (id and name) and all raised objects within that room.

Therefore, altered the data model:
A room now lists the connected users as HashMap id->name rather than only the
ids. This is to provide the list of connected users with both id and name.